Added Game class and quit game functionality
Game class is needed to separate the logic for running the game from main method.
Main method will create a new Game object and call Game.startGame() to make the
game run. Quit game feature is added. Without quit game feature, a user can't
exit the game prematurely and is forced to play all the way to the end or close
the app running the game, which is a bad experience. Any time user input is
required, user should enter Q if they would like to quit the game. All new
features which take user input should check for Q from the user and ensure
control passes to endGame() immediately.