Added method for user and computer to attack.
Per the requirements, user and computer taking turns attacking. A valid attack is a coordinate which is on the board, and which wasn't previously guessed. If the coordinate you attacked contains a ship, the ship is sunk, and the map is updated. Attacking one's own ship is a valid move, per the requirements. This completes the game, per the requirements.