Refactors Game class to take user input in 'X, Y' format.
Previously, users were prompted to enter an X coordinate, then
prompted a second time for the Y coordinate. The changes allow
the user to enter 'X, Y' or 'X,Y' on one line, and uses regular
expressions to validate the input. The logic in the Game class
has also been extensively refactored to improve readability and
maintainability by separating more of the code into helper methods.